% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RILStEp.R
\name{check_epistasis}
\alias{check_epistasis}
\title{Calculate BayesFactor for Comparison between Model1 (without Epistasis) vs Model2 (with Epistasis)}
\usage{
check_epistasis(
  pair_index,
  genotype_phenotype_dataset,
  phenotype_name,
  peak_qtls,
  pairs
)
}
\arguments{
\item{pair_index}{indices of SNP pair to check epistasis}

\item{genotype_phenotype_dataset}{dataset of genotype and phenotype}

\item{phenotype_name}{trait's name}

\item{peak_qtls}{QTL-like SNP index list from extract_peak_qtls() or user specify}

\item{pairs}{SNP pairs}
}
\description{
Calculate BayesFactor for Comparison between Model1 (without Epistasis) vs Model2 (with Epistasis)
}
