% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RILStEp.R
\name{load_data}
\alias{load_data}
\title{load dataset to use GWAS and RILStEp}
\usage{
load_data(phenotype_path, genotype_path, phenotype_name)
}
\arguments{
\item{phenotype_path}{path to csv file of phenotype data}

\item{genotype_path}{path to csv file of genotype data}

\item{phenotype_name}{trait's name}
}
\description{
load dataset to use GWAS and RILStEp
}
