% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RILStEp.R
\name{marker2pos}
\alias{marker2pos}
\title{Convert Marker number to position data (chrXX_ZZZZZ)}
\usage{
marker2pos(marker, marker_data)
}
\arguments{
\item{marker}{marker index of SNP}

\item{marker_data}{indices of all SNP positions}
}
\description{
Convert Marker number to position data (chrXX_ZZZZZ)
}
