% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RILStEp.R
\name{pos2marker}
\alias{pos2marker}
\title{Convert position data (chrXX_ZZZZZ) to Marker number}
\usage{
pos2marker(position, marker_data)
}
\arguments{
\item{position}{position data of SNP}

\item{marker_data}{indices of all SNP positions}
}
\description{
Convert position data (chrXX_ZZZZZ) to Marker number
}
