% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RILStEp.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{draw heatmap of bayes factor}
\usage{
draw_heatmap(result, filename, marker_data)
}
\arguments{
\item{result}{bayesfactor of all combination of SNPs}

\item{filename}{name of save file}

\item{marker_data}{indices of all SNP positions}
}
\description{
draw heatmap of bayes factor
}
