% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RILStEp.R
\name{extract_peak_qtls}
\alias{extract_peak_qtls}
\title{extracts QTL-like SNPs based on GWASpoly package}
\usage{
extract_peak_qtls(data_for_gwas, phenotype_name, output, core_num)
}
\arguments{
\item{data_for_gwas}{phenotype and genotype data for GWASpoly from load_data}

\item{phenotype_name}{trait's name}

\item{output}{prefix of output files}

\item{core_num}{the number of CPU cores to use}
}
\description{
extracts QTL-like SNPs based on GWASpoly package
}
